+++ jest.config.js
@@ -8,7 +8,6 @@ module.exports = {
   testPathIgnorePatterns: ["<rootDir>/.next/", "<rootDir>/node_modules/"],
   moduleNameMapper: {
     "\\.(css)$": "<rootDir>/src/__mocks__/styleMock.js",
-    "path.json": "<rootDir>/src/__mocks__/pathMock.json",
     "schema.js": "<rootDir>/src/__mocks__/schemaMock.js",
     "logging.service": "<rootDir>/src/__mocks__/logging.service.js",
     "statusEmitter.service": "<rootDir>/src/__mocks__/statusEmitter.service.js"