+++ pages/operator-company-details.js
@@ -62,22 +62,24 @@ const LimitedCompanyDetails = props => (
         </InputField>
       </ContentItem.B_30_15>
 
-      <HiddenText summaryText={"I don't know my Companies House number"}>
-        {/* TODO JMB: replace the span with a paragraph once it's possible to pass an array or similar to Paragraph for the link */}
-        <span>
-          Every limited company has to be registered with Companies House, and
-          will have a company reference number. You can find your company
-          reference number by visiting the{" "}
-          <a
-            href="https://beta.companieshouse.gov.uk/"
-            target="_blank"
-            rel="noopener noreferrer"
-            id="link-companies-house"
-          >
-            Companies House website
-          </a>.
-        </span>
-      </HiddenText>
+      <ContentItem.B_30_15>
+        <HiddenText summaryText={"I don't know my Companies House number"}>
+          {/* TODO JMB: replace the span with a paragraph once it's possible to pass an array or similar to Paragraph for the link */}
+          <span>
+            Every limited company has to be registered with Companies House, and
+            will have a company reference number. You can find your company
+            reference number by visiting the{" "}
+            <a
+              href="https://beta.companieshouse.gov.uk/"
+              target="_blank"
+              rel="noopener noreferrer"
+              id="link-companies-house"
+            >
+              Companies House website
+            </a>.
+          </span>
+        </HiddenText>
+      </ContentItem.B_30_15>
 
       <ContinueButton editMode={props.editMode} />
     </form>