+++ pages/registration-summary.js
@@ -12,12 +12,15 @@ import PropTypes from "prop-types";
 const RegistrationSummary = props => (
   <FsaLayout {...props}>
     <BackButton editMode={props.editMode} originator="registration-summary" />
+
     <ContentItem.B_30_15>
-      <Header level={2}>Check your answers</Header>
-      <HintText>You must check your answers before you continue</HintText>
-    </ContentItem.B_30_15>
-    <ContentItem.B_30_15>
-      <SummaryTable {...props.transformedData} />
+      <ContentItem.B_30_15>
+        <Header level={2}>Check your answers</Header>
+        <HintText>You must check your answers before you continue</HintText>
+      </ContentItem.B_30_15>
+      <ContentItem.B_30_15>
+        <SummaryTable {...props.transformedData} />
+      </ContentItem.B_30_15>
     </ContentItem.B_30_15>
 
     <form action="/continue/registration-summary" method="post">