+++ src/__tests__/SummaryTable.test.js
@@ -27,7 +27,8 @@ const mandatoryTableRows = [
   "contactRepresentativeRow",
   "establishmentOpeningDateRow",
   "businessTypeRow",
-  "businessOtherDetailsRow"
+  "businessOtherDetailsRow",
+  "establishmentOpeningDaysRow"
 ];
 
 const editableTableRows = [
@@ -79,7 +80,8 @@ const testMandatoryAnswers = {
   establishment_opening_date: "2018-12-06",
   establishment_type: "Mobile or moveable premise",
   business_type: "Livestock farm",
-  business_other_details: "This is the best business in the world"
+  business_other_details: "This is the best business in the world",
+  opening_day_monday: "Monday"
 };
 
 // a supplementary set of all optional answer fields with example data
@@ -144,9 +146,11 @@ describe("<SummaryTable />", () => {
     it("the number of table rows matches the allTableRows array", () => {
       const rows = wrapperComprehensive
         .find("Row")
-        .findWhere(row => row.prop("className") !== "TITLE")
+        .findWhere(row => {
+          const classNameString = row.prop("className") || "";
+          return classNameString.includes("TITLE") === false;
+        })
         .find("Row");
-
       expect(rows.length).toEqual(allTableRows.length);
     });
 
@@ -207,15 +211,6 @@ describe("<SummaryTable />", () => {
           expect(element.text()).toBe("");
         }
       }
-    });
-  });
-
-  describe("when answers are missing", () => {
-    it("does not render any table rows", () => {
-      allTableRows.forEach(tableRowName => {
-        const row = wrapperMissing.find(`Row#${tableRowName}`);
-        expect(row.length).toBe(0);
-      });
     });
   });
 });