+++ src/components/SummaryTable.js
@@ -28,22 +28,30 @@ const StyledTableRow = styled("div")({
   color: `${COLOUR.GREY_1}`
 });
 
+const FormattedTableRow = styled(Table.Row)`
+  display: grid;
+  ${props =>
+    props.acPage
+      ? "grid-template-columns: 1fr 1fr;"
+      : "grid-template-columns: 1fr 1fr 70px;"};
+`;
+
 const AnchorTag = asAnchor("a");
 
 const OperatorDetailsTable = props => (
   <React.Fragment>
     {props.applicationCompletePage ? (
-      <Table.Row>
+      <FormattedTableRow acPage={props.applicationCompletePage}>
         <Table.CellHeader>
           <div />
         </Table.CellHeader>
         <Table.Cell>
           <div />
         </Table.Cell>
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
-    <Table.Row className="TITLE">
+    <FormattedTableRow acPage={props.applicationCompletePage} className="TITLE">
       <Table.CellHeader>
         <Paragraph mb={0}>**Operator details**</Paragraph>
       </Table.CellHeader>
@@ -55,10 +63,13 @@ const OperatorDetailsTable = props => (
           <div />
         </Table.Cell>
       )}
-    </Table.Row>
+    </FormattedTableRow>
 
     {props.operator_type ? (
-      <Table.Row id="operatorTypeRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorTypeRow"
+      >
         <Table.CellHeader>Operator type</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="operator_type">
@@ -70,11 +81,14 @@ const OperatorDetailsTable = props => (
             <div />
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_company_name ? (
-      <Table.Row id="operatorCompanyNameRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorCompanyNameRow"
+      >
         <Table.CellHeader>Company name</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div id="operator_company_name" className="bold">
@@ -91,11 +105,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_company_house_number ? (
-      <Table.Row id="operatorCompaniesHouseRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorCompaniesHouseRow"
+      >
         <Table.CellHeader>Company number</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div id="operator_company_house_number" className="bold">
@@ -112,11 +129,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_charity_name ? (
-      <Table.Row id="operatorCharityNameRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorCharityNameRow"
+      >
         <Table.CellHeader>Charity name</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div id="operator_charity_name" className="bold">
@@ -133,11 +153,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_charity_number ? (
-      <Table.Row id="operatorCharityNumberRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorCharityNumberRow"
+      >
         <Table.CellHeader>Charity number</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div id="operator_charity_number" className="bold">
@@ -154,11 +177,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_first_name ? (
-      <Table.Row id="operatorNameRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorNameRow"
+      >
         <Table.CellHeader>Name</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold">
@@ -173,11 +199,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_first_line ? (
-      <Table.Row id="operatorAddressRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorAddressRow"
+      >
         <Table.CellHeader>Operator address</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold">
@@ -192,11 +221,14 @@ const OperatorDetailsTable = props => (
             <div />
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_primary_number ? (
-      <Table.Row id="operatorContactDetailsRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorContactDetailsRow"
+      >
         <Table.CellHeader>Phone number</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold">
@@ -218,11 +250,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.operator_email ? (
-      <Table.Row id="operatorEmailRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="operatorEmailRow"
+      >
         <Table.CellHeader>Email address</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="operator_email">
@@ -239,11 +274,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.contact_representative_email ? (
-      <Table.Row id="contactRepresentativeRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="contactRepresentativeRow"
+      >
         <Table.CellHeader>Designated contact</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <StyledTableRow>
@@ -275,14 +313,14 @@ const OperatorDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
   </React.Fragment>
 );
 
 const EstablishmentDetailsTable = props => (
   <React.Fragment>
-    <Table.Row className="TITLE">
+    <FormattedTableRow acPage={props.applicationCompletePage} className="TITLE">
       <Table.CellHeader>
         <Paragraph mb={0}>**Establishment details**</Paragraph>
       </Table.CellHeader>
@@ -294,10 +332,13 @@ const EstablishmentDetailsTable = props => (
           <div />
         </Table.Cell>
       )}
-    </Table.Row>
+    </FormattedTableRow>
 
     {props.establishment_trading_name ? (
-      <Table.Row id="establishmentTradingNameRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="establishmentTradingNameRow"
+      >
         <Table.CellHeader>Trading name</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="establishment_trading_name">
@@ -314,11 +355,14 @@ const EstablishmentDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.establishment_first_line ? (
-      <Table.Row id="establishmentAddressRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="establishmentAddressRow"
+      >
         <Table.CellHeader>Establishment address</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold">
@@ -341,11 +385,14 @@ const EstablishmentDetailsTable = props => (
             <div />
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.establishment_type ? (
-      <Table.Row id="establishmentAddressTypeRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="establishmentAddressTypeRow"
+      >
         <Table.CellHeader>Address type</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="establishment_type">
@@ -362,11 +409,14 @@ const EstablishmentDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.establishment_primary_number ? (
-      <Table.Row id="establishmentContactDetailsRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="establishmentContactDetailsRow"
+      >
         <Table.CellHeader>Phone number</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold">
@@ -388,11 +438,14 @@ const EstablishmentDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.establishment_email ? (
-      <Table.Row id="establishmentEmailRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="establishmentEmailRow"
+      >
         <Table.CellHeader>Email address</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="establishment_email">
@@ -409,11 +462,14 @@ const EstablishmentDetailsTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.establishment_opening_date ? (
-      <Table.Row id="establishmentOpeningDateRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="establishmentOpeningDateRow"
+      >
         <Table.CellHeader>Trading date</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="establishment_opening_date">
@@ -425,14 +481,52 @@ const EstablishmentDetailsTable = props => (
             <div />
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
+
+    <FormattedTableRow
+      acPage={props.applicationCompletePage}
+      id="establishmentOpeningDaysRow"
+    >
+      <Table.CellHeader>Opening days</Table.CellHeader>
+      <Table.Cell className="summaryTableDataCell">
+        <div className="bold">
+          <div id="opening_days_irregular">
+            {props.opening_days_irregular ||
+              props.open_some_days_summary_table || (
+                <div>
+                  <div id="opening_day_monday">{props.opening_day_monday}</div>
+                  <div id="opening_day_tuesday">
+                    {props.opening_day_tuesday}
+                  </div>
+                  <div id="opening_day_wednesday">
+                    {props.opening_day_wednesday}
+                  </div>
+                  <div id="opening_day_thursday">
+                    {props.opening_day_thursday}
+                  </div>
+                  <div id="opening_day_friday">{props.opening_day_friday}</div>
+                  <div id="opening_day_saturday">
+                    {props.opening_day_saturday}
+                  </div>
+                  <div id="opening_day_sunday">{props.opening_day_sunday}</div>
+                </div>
+              )}
+          </div>
+        </div>
+      </Table.Cell>
+      {props.applicationCompletePage ? null : (
+        <Table.Cell>
+          <div />
+        </Table.Cell>
+      )}
+    </FormattedTableRow>
   </React.Fragment>
 );
 
 const FoodActivitiesTable = props => (
   <React.Fragment>
-    <Table.Row className="TITLE">
+    <FormattedTableRow acPage={props.applicationCompletePage} className="TITLE">
       <Table.CellHeader>
         <Paragraph mb={0}>**Activities**</Paragraph>
       </Table.CellHeader>
@@ -444,10 +538,13 @@ const FoodActivitiesTable = props => (
           <div />
         </Table.Cell>
       )}
-    </Table.Row>
+    </FormattedTableRow>
 
     {props.customer_type ? (
-      <Table.Row id="activitiesCustomersRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="activitiesCustomersRow"
+      >
         <Table.CellHeader>Customers</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="customer_type">
@@ -464,11 +561,14 @@ const FoodActivitiesTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.business_type ? (
-      <Table.Row id="businessTypeRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="businessTypeRow"
+      >
         <Table.CellHeader>Business type</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="business_type">
@@ -491,11 +591,14 @@ const FoodActivitiesTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.import_export_activities ? (
-      <Table.Row id="activitiesBusinessImportExportRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="activitiesBusinessImportExportRow"
+      >
         <Table.CellHeader>Import and export</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="import_export_activities">
@@ -512,11 +615,14 @@ const FoodActivitiesTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
 
     {props.business_other_details ? (
-      <Table.Row id="businessOtherDetailsRow">
+      <FormattedTableRow
+        acPage={props.applicationCompletePage}
+        id="businessOtherDetailsRow"
+      >
         <Table.CellHeader>Additional details</Table.CellHeader>
         <Table.Cell className="summaryTableDataCell">
           <div className="bold" id="business_other_details">
@@ -533,48 +639,57 @@ const FoodActivitiesTable = props => (
             </AnchorTag>
           </Table.Cell>
         )}
-      </Table.Row>
+      </FormattedTableRow>
     ) : null}
   </React.Fragment>
 );
 
 const DeclarationTable = props => (
   <React.Fragment>
-    <Table.Row className="TITLE">
+    <FormattedTableRow acPage={props.applicationCompletePage} className="TITLE">
       <Table.CellHeader>
         <Paragraph mb={0}>**Declaration**</Paragraph>
       </Table.CellHeader>
       <Table.Cell>
         <div />
       </Table.Cell>
-    </Table.Row>
+    </FormattedTableRow>
 
-    <Table.Row id="declaration1Row">
+    <FormattedTableRow
+      acPage={props.applicationCompletePage}
+      id="declaration1Row"
+    >
       <Table.CellHeader>
         <span id="declaration1">{props.declaration1}</span>
       </Table.CellHeader>
       <Table.Cell className="summaryTableDataCell">
         <div className="bold">Accepted</div>
       </Table.Cell>
-    </Table.Row>
+    </FormattedTableRow>
 
-    <Table.Row id="declaration2Row">
+    <FormattedTableRow
+      acPage={props.applicationCompletePage}
+      id="declaration2Row"
+    >
       <Table.CellHeader>
         <span id="declaration2">{props.declaration2}</span>
       </Table.CellHeader>
       <Table.Cell className="summaryTableDataCell">
         <div className="bold">Accepted</div>
       </Table.Cell>
-    </Table.Row>
+    </FormattedTableRow>
 
-    <Table.Row id="declaration3Row">
+    <FormattedTableRow
+      acPage={props.applicationCompletePage}
+      id="declaration3Row"
+    >
       <Table.CellHeader>
         <span id="declaration3">{props.declaration3}</span>
       </Table.CellHeader>
       <Table.Cell className="summaryTableDataCell">
         <div className="bold">Accepted</div>
       </Table.Cell>
-    </Table.Row>
+    </FormattedTableRow>
   </React.Fragment>
 );